refactor: moves HybridModuleDetector and VendorModuleDetector to Classes directory
